Speedup of the profile identifying process.
The speed of the process identifying individual firn profiles in the
dataset is enhanced. It is now based on building hashes over the
metadata of every data point summarized in a string.

Some tests show the result is the same as before. Even when the order
the profiles are numbered has changed.